Revert "Feature suh"

+++ src/screen/ScoreScreen.java
@@ -30,7 +30,7 @@ public class ScoreScreen extends Screen {
 	/** Current score. */
 	private int score;
 	/** Player lives left. */
-	private double livesRemaining;
+	private int livesRemaining;
 	/** Total bullets shot by the player. */
 	private int bulletsShot;
 	/** Total ships destroyed by the player. */
